Exception handling

+++ Multiserver/src/activitystreamer/util/MapComparator.java
@@ -14,10 +14,10 @@ public class MapComparator implements Comparator<String> {
     }
     @Override
     public int compare(String o1, String o2) {
-        int level1 = (Integer) base.get(o1).get("level");
-        int level2 = (Integer) base.get(o2).get("level");
-        int rank1 = (Integer) base.get(o1).get("rank");
-        int rank2 = (Integer) base.get(o2).get("rank");
+        int level1 = ((Long) base.get(o1).get("level")).intValue();
+        int level2 = ((Long) base.get(o2).get("level")).intValue();
+        int rank1 = ((Long) base.get(o1).get("rank")).intValue();
+        int rank2 = ((Long) base.get(o2).get("rank")).intValue();
         if(level1 > level2){
             return 1;
         }else if(level1 < level2){
